Add destination endpoints, mappers, and DTOs
Introduce admin endpoints and backing plumbing for creating and deleting destinations. AdminController now exposes POST /destinations, DELETE /destinations/{id} and keeps GET /fetch-from-api; requests are validated with Jakarta annotations. Add DestinationRequest (with validation) and MessageResponse DTOs, and a DestinationMapper to convert between DTOs and the Destination entity. DestinationService interface extended with addDestination and deleteDestination; DestinationServiceImpl implements these using DestinationRepository and DestinationMapper, checks for duplicates on add and throws EntityNotFound on delete. Change DestinationRepository ID type to Long. Country mapping updated: CountryResponse is built via Lombok @Builder and CountryMapper simplified/cleaned up. Also add Lombok @RequiredArgsConstructor and remove some comment clutter and minor formatting tweaks.

+++ src/main/java/com/example/travel_planner/repository/DestinationRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DestinationRepository extends JpaRepository<Destination, Integer> {
+public interface DestinationRepository extends JpaRepository<Destination, Long> {
 
     @Query("SELECT d FROM Destination d WHERE d.approved = true")
     List<Destination> findByApprovedTrue();
